up strategy

+++ src/factory/simple/OperationDiv.java
@@ -6,7 +6,7 @@ public class OperationDiv extends Operation {
     private double n2 = this.getN2();
     @Override
     public double result() {
-        if (n2 == 0) {
+        if ((int)n2 == 0) {
             throw new RuntimeException("divided by 0");
         }
         return n1 / n2;
